refactor(jobs): split WorkerLoopService into dispatcher/poller/loop (max-params 2/3)
WorkerLoopService injected 7 deps (config, claim, lease, apiKeys, now,
workerPool, metrics) — `now` (WORKER_LOOP_NOW) was a never-provided dead
seam. Three-way split:
- JobDispatcherService (claim, workerPool, metrics) — runs one claim to
  completion: renew/cancel poll, handler/worker-pool invoke, OTel span,
  terminal complete/fail/cancelled. Takes the shutdown AbortSignal in.
- WorkerPollerService (dispatcher, claim, apiKeys) — per-jobType poll
  loop with weighted-fair tenant ordering + recent-claim decay; cadences
  via env.
- WorkerLoopService (poller, lease, metrics) — leader election, handler
  registry, lifecycle; enabled/renew via env. Owns the abortController,
  passes its signal to the poller via PollControl.
Shared types (JobContext/JobHandler/RegisteredHandler) moved to
worker-loop.types.ts (re-exported for compat). Each class ≤3 deps;
register()/registeredTypes()/leader() surface unchanged. Specs split
across dispatcher/poller/loop.

Part of the max-params=3 god-class split program.
