+++ src/jobs/jobs.module.ts
@@ -4,6 +4,8 @@ import { JobRunService } from './job-run.service';
 import { JobClaimService } from './job-claim.service';
 import { LeaderLeaseService } from './leader-lease.service';
 import { WorkerLoopService } from './worker-loop.service';
+import { WorkerPollerService } from './worker-poller.service';
+import { JobDispatcherService } from './job-dispatcher.service';
 import { LeaseManagerService } from './lease-manager.service';
 import { JobReaperService } from './job-reaper.service';
 import { JobWorkerPool } from './job-worker-pool.service';
@@ -34,6 +36,8 @@ import { DistributedLeaseGuard } from '../common/distributed-lease.guard';
     JobRunService,
     JobClaimService,
     LeaderLeaseService,
+    JobDispatcherService,
+    WorkerPollerService,
     WorkerLoopService,
     LeaseManagerService,
     JobReaperService,
